backend/requirements.txt
Flask
flask-cors
python-dotenv
requests
PyJWT
cryptography
SQLAlchemy
gunicorn